--- a/modules/sfp__stor_db.py
+++ b/modules/sfp__stor_db.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-#
 # -------------------------------------------------------------------------------
 # Name:         sfp_stor_db
 # Purpose:      SpiderFoot plug-in for storing events to the local SpiderFoot
@@ -11,8 +10,7 @@
 # Copyright:   (c) Steve Micallef 2012
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-#
-from __future__ import absolute_import
+
 from sflib import SpiderFoot, SpiderFootPlugin