--- a/modules/sfp_adblock.py
+++ b/modules/sfp_adblock.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+#
 # -------------------------------------------------------------------------------
 # Name:         sfp_adblock
 # Purpose:      SpiderFoot plug-in to test if external/internally linked pages
@@ -10,7 +11,8 @@
 # Copyright:   (c) Steve Micallef 2014
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-
+#
+from __future__ import absolute_import
 import adblockparser
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent