--- a/modules/sfp_bingsearch.py
+++ b/modules/sfp_bingsearch.py
@@ -30,7 +30,7 @@ class sfp_bingsearch(SpiderFootPlugin):
 
     results = list()
 
-    def setup(self, sfc, userOpts=dict()):
+    def setup(self, sfc, userOpts={}):
         self.sf = sfc
         self.results = list()