--- a/modules/sfp_blacklist.py
+++ b/modules/sfp_blacklist.py
@@ -153,7 +153,7 @@ class sfp_blacklist(SpiderFootPlugin):
         srcModuleName = event.module
         eventData = event.data
         parentEvent = event
-        addrlist = list()
+        addrlist = []
 
         self.sf.debug("Received event, " + eventName + ", from " + srcModuleName)