--- a/modules/sfp_cookie.py
+++ b/modules/sfp_cookie.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-#
 # -------------------------------------------------------------------------------
 # Name:         sfp_cookie
 # Purpose:      SpiderFoot plug-in for extracting cookies from HTTP headers.
@@ -10,8 +9,7 @@
 # Copyright:   (c) Steve Micallef 2014
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-#
-from __future__ import absolute_import
+
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent