--- a/modules/sfp_crossref.py
+++ b/modules/sfp_crossref.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+#
 # -------------------------------------------------------------------------------
 # Name:         sfp_crossref
 # Purpose:      SpiderFoot plug-in for scanning links identified from the
@@ -12,7 +13,8 @@
 # Copyright:   (c) Steve Micallef 2012
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-
+#
+from __future__ import absolute_import
 import re
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent