--- a/modules/sfp_defaced.py
+++ b/modules/sfp_defaced.py
@@ -38,7 +38,7 @@ class sfp_defaced(SpiderFootPlugin):
 
     results = list()
 
-    def setup(self, sfc, userOpts=dict()):
+    def setup(self, sfc, userOpts={}):
         self.sf = sfc
         self.results = list()