--- a/modules/sfp_defaced.py
+++ b/modules/sfp_defaced.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-#
 # -------------------------------------------------------------------------------
 # Name:         sfp_defaced
 # Purpose:      Checks if a domain or IP appears on the zone-h.org defacement
@@ -11,8 +10,7 @@
 # Copyright:   (c) Steve Micallef, 2014
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-#
-from __future__ import absolute_import
+
 import time
 import re
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent