--- a/modules/sfp_email.py
+++ b/modules/sfp_email.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+#
 # -------------------------------------------------------------------------------
 # Name:         sfp_email
 # Purpose:      SpiderFoot plug-in for scanning retreived content by other
@@ -10,7 +11,8 @@
 # Copyright:   (c) Steve Micallef 2012
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-
+#
+from __future__ import absolute_import
 import re
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent