--- a/modules/sfp_errors.py
+++ b/modules/sfp_errors.py
@@ -40,11 +40,11 @@ class sfp_errors(SpiderFootPlugin):
     }
 
     # Target
-    results = dict()
+    results = {}
 
-    def setup(self, sfc, userOpts=dict()):
+    def setup(self, sfc, userOpts={}):
         self.sf = sfc
-        self.results = dict()
+        self.results = {}
 
         for opt in userOpts.keys():
             self.opts[opt] = userOpts[opt]