--- a/modules/sfp_errors.py
+++ b/modules/sfp_errors.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+#
 # -------------------------------------------------------------------------------
 # Name:         sfp_errors
 # Purpose:      Identify common error messages in content like SQL errors, etc.
@@ -9,7 +10,8 @@
 # Copyright:   (c) Steve Micallef 2015
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-
+#
+from __future__ import absolute_import
 import re
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent