--- a/modules/sfp_errors.py
+++ b/modules/sfp_errors.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-#
 # -------------------------------------------------------------------------------
 # Name:         sfp_errors
 # Purpose:      Identify common error messages in content like SQL errors, etc.
@@ -10,8 +9,7 @@
 # Copyright:   (c) Steve Micallef 2015
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-#
-from __future__ import absolute_import
+
 import re
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent