--- a/modules/sfp_filemeta.py
+++ b/modules/sfp_filemeta.py
@@ -37,7 +37,7 @@ class sfp_filemeta(SpiderFootPlugin):
 
     results = list()
 
-    def setup(self, sfc, userOpts=dict()):
+    def setup(self, sfc, userOpts={}):
         self.sf = sfc
         self.results = list()