--- a/modules/sfp_filemeta.py
+++ b/modules/sfp_filemeta.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-#
 # -------------------------------------------------------------------------------
 # Name:         sfp_filemeta
 # Purpose:      From Spidering and from searching search engines, extracts file
@@ -11,8 +10,7 @@
 # Copyright:   (c) Steve Micallef 2014
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-#
-from __future__ import absolute_import
+
 import mimetypes
 import metapdf
 import openxmllib