--- a/modules/sfp_geoip.py
+++ b/modules/sfp_geoip.py
@@ -20,11 +20,11 @@ class sfp_geoip(SpiderFootPlugin):
 
     # Default options
     opts = {}
-    results = dict()
+    results = {}
 
-    def setup(self, sfc, userOpts=dict()):
+    def setup(self, sfc, userOpts={}):
         self.sf = sfc
-        self.results = dict()
+        self.results = {}
 
         for opt in userOpts.keys():
             self.opts[opt] = userOpts[opt]