--- a/modules/sfp_geoip.py
+++ b/modules/sfp_geoip.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-#
 # -------------------------------------------------------------------------------
 # Name:         sfp_geoip
 # Purpose:      SpiderFoot plug-in to identify the Geo-location of IP addresses
@@ -11,8 +10,7 @@
 # Copyright:   (c) Steve Micallef 2013
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-#
-from __future__ import absolute_import
+
 import json
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent