--- a/modules/sfp_googlesearch.py
+++ b/modules/sfp_googlesearch.py
@@ -31,7 +31,7 @@ class sfp_googlesearch(SpiderFootPlugin):
     # Target
     results = list()
 
-    def setup(self, sfc, userOpts=dict()):
+    def setup(self, sfc, userOpts={}):
         self.sf = sfc
         self.results = list()