--- a/modules/sfp_googlesearch.py
+++ b/modules/sfp_googlesearch.py
@@ -29,11 +29,11 @@ class sfp_googlesearch(SpiderFootPlugin):
     }
 
     # Target
-    results = list()
+    results = []
 
     def setup(self, sfc, userOpts={}):
         self.sf = sfc
-        self.results = list()
+        self.results = []
 
         for opt in userOpts.keys():
             self.opts[opt] = userOpts[opt]