--- a/modules/sfp_googlesearch.py
+++ b/modules/sfp_googlesearch.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+#
 # -------------------------------------------------------------------------------
 # Name:         sfp_googlesearch
 # Purpose:      Searches Google for content related to the domain in question.
@@ -9,7 +10,8 @@
 # Copyright:   (c) Steve Micallef 2012
 # Licence:     GPL
 # -------------------------------------------------------------------------------
-
+#
+from __future__ import absolute_import
 from sflib import SpiderFoot, SpiderFootPlugin, SpiderFootEvent