--- a/modules/sfp_intfiles.py
+++ b/modules/sfp_intfiles.py
@@ -37,7 +37,7 @@ class sfp_intfiles(SpiderFootPlugin):
 
     results = list()
 
-    def setup(self, sfc, userOpts=dict()):
+    def setup(self, sfc, userOpts={}):
         self.sf = sfc
         self.results = list()