Simplifies a few things, fixes more PEP violations, and improves performance a little, and the addition of __future__ imports.

Simplifying assignment:
* Augmented Assignment: 4d0c43e250e1118a895edefe95ccadd462f295fe

PEP Violation Fixes:
* PEP 8 – Comparison to 'None': 7bc2ddc7266e159c912636ffaa0a28edab60ab39
* PEP 263 – File Encoding: f18a50d1642fcc7c85f71350316a9af770ea7841
* More PEP Violation Fixes: 297ece1bab4d55d7fdc920ab04bbc886799c2849

Replaced empty Dictionary 'dict()'and List 'list()' creation with '{}' and '[]', as this is far more efficient:
* Dictionary '{}' Instead of 'dict()': fd33bc2d7aac76b256fc364025cf16cd3d4f9a12
```
python2 -m timeit -n 1000000 -r 5 -v 'dict()' 
raw times: 0.137 0.0916 0.0806 0.0807 0.0806
1000000 loops, best of 5: 0.0806 usec per loop

python2 -m timeit -n 1000000 -r 5 -v '{}' 
raw times: 0.0517 0.0419 0.0362 0.0323 0.0295
1000000 loops, best of 5: 0.0295 usec per loop
```
* List '[]' Instead of 'list()': 353d27f1c593ceb7ba49c63039529a9ac113c659
```
python2 -m timeit -n 1000000 -r 5 -v 'list()'
raw times: 0.141 0.0933 0.0897 0.0897 0.0899
1000000 loops, best of 5: 0.0897 usec per loop

python2 -m timeit -n 1000000 -r 5 -v '[]'    
raw times: 0.0504 0.0407 0.0353 0.0316 0.0288
1000000 loops, best of 5: 0.0288 usec per loop
```

For some unfathomable reason Git is acting like I modified the entire file, my assumption is that the line ending was converted from the Windows '\r\n' to the Unix '\n'.  Ultimately this doesn't matter, but I wanted to assure you that I indeed only added a few __future__ imports and converted the 'print' statements to Python 3 syntax to ease the transition to Python 3 in the future.  Unfortunately, I would have to blow away all of my changes to resolve what Git did.:
* __future__ for Python 3 Transition: 46fdf01a1e69aff045e55e60e3579cb40003a9fa
* https://docs.python.org/2/library/__future__.html
* __NOTE!: This change was rolled back with 46fdf01a1e69aff045e55e60e3579cb40003a9fa as 'modernize' modified the line-endings, rendering changes indistinguishable from unchanged code in the diff.__

Misc Changes:
* Rewrite as Dictionary Literal: b55160cd9e34a4e3a3c62291d74185189b3091a6
* Fix Classic Style Class Usage: b0a2cfd1f4da8f9fb63a08817d2341d32f7dd17f