@smicallef 2015-04-08
This is brilliant, but a couple of questions - are you using a tool to do this (I hope you aren't doing it manually for the sake of your sanity) and based on the diff basically every line is changing, but it's hard for me to see exactly what.