@Gormogon 2015-04-08
99% of what I've been doing is manual, as automatic tools tend to f*ck up the code.  I tried to use python-modernize to quickly convert the print statements to Python 3 style and add the __future__ import statements, however this appears to be what changed the line endings and caused the diff to go crazy.

I just created a 'revert' commit to undo 46fdf01a1e69aff045e55e60e3579cb40003a9fa as well as a second commit to quickly resolve conflicts.

If the addition of __future__ import statements and the conversion of print statements is desired, I will gladly make these changes at a later date and make sure to fix the line endings before commit so this doesn't happen again.